+++ projects/komponentkartan/src/assets/komponentkartan.scss
@@ -59,3 +59,4 @@
 @import "partials/_components.ring-with-text.scss";
 @import "partials/_components.toggle-button.scss";
 @import "partials/_components.search-result.scss";
+@import "partials/components.pagination";
