+++ projects/komponentkartan/src/assets/partials/_components.list.scss
@@ -93,71 +93,6 @@ vgr-list.list--new-item-added {
   background-color:white; 
   padding-top: 10px; 
   padding-bottom:10px; 
-  .list__footer__pagination {
-    display: flex;
-    justify-content: center;
-    .list__footer__pagination__page-button {         
-      position: relative;
-      box-sizing: border-box;
-      padding: 0;
-      border: 0;
-      border-radius: 1px;
-      vertical-align: baseline;
-      text-align: center;
-      cursor: pointer;
-      @extend .color--primary;
-      background-color: transparent;
-      border-color: transparent;
-      border-width: 1px;
-      border-style: solid;
-      font-size: $font-size--small;
-      font-family: $font-family;
-      line-height: 28px;
-      min-width: 10px;
-      padding-left: 5px;
-      padding-right: 5px;
-      margin-left: 7px;
-      margin-right: 7px;
-      &:active {
-        transform: scale(0.96);
-        outline: 0;
-      }
-      &:hover:not(.page-active),
-      &:focus:not(.page-active) {
-        @extend .background-color--primary-dimmed;
-        border-color: $primary-color--default-dimmed;
-        @extend .color--primary;
-      } 
-      &:focus:not(:active) {
-        outline: $focus-outline-width solid $focus-color;
-      }
-      &.page-active {
-        @extend .background-color--primary;
-        color: $text-color--light;
-      }  
-      &.previous {
-        margin-right: 33px;
-        & > span:first-child {
-          font-weight: bold;
-          padding-right: 8px;
-        }
-      }
-      &.next {
-        margin-left: 33px;
-        & > span:last-child {
-          font-weight: bold;
-          padding-left: 8px;
-        }
-      }    
-    }
-    .list__footer__pagination__dots {
-      font-size: $font-size--small;
-      font-family: $font-family;
-      line-height: 30px;
-      margin-left: 10px;
-      margin-right: 10px;
-    }
-  }
 }
 
 .list__items {
