+++ projects/komponentkartan/src/lib/controls/list/list.component.html
@@ -10,14 +10,5 @@
   <ng-content select="vgr-list-item"></ng-content>
 </div>
 <div class="list__footer" *ngIf="pages > 1">
-  <div class="list__footer__pagination">
-    <div *ngFor="let item of pageItems; first as isPrevious; last as isNext;">
-        <button *ngIf="item.label !== '...'" #pageButton [type]="button" class="list__footer__pagination__page-button" [ngClass]="{'page-active' : item.active, 'previous' : isPrevious, 'next' : isNext }" (click)="item.action()" (focus)="onPageButtonFocus($event)" (blur)="onPageButtonBlur($event)">
-          <span *ngIf="isPrevious">&lt;</span>
-          <span>{{ item.label }}</span>
-          <span *ngIf="isNext" >&gt;</span>
-        </button>
-        <span class="noselect list__footer__pagination__dots" *ngIf="item.label === '...'">{{ item.label }}</span>
-    </div>
-  </div>
+  <vgr-pagination [pages]="pages" [activePage]="activePage" (pageChanged)="onPageChanged($event)"></vgr-pagination>
 </div>
