+++ projects/komponentkartan/src/lib/controls/pagination/page-item.ts
@@ -1,0 +1,5 @@
+export interface PageItem {
+    label: string;
+    active: boolean;
+    action: () => void;
+}
